+++ test/mix/tasks/pleroma/instance_test.exs
@@ -69,7 +69,7 @@ defmodule Mix.Tasks.Pleroma.InstanceTest do
         "test/uploads",
         "--static-dir",
         "./test/../test/instance/static/",
-        "--strip-uploads-location",
+        "--strip-uploads-metadata",
         "y",
         "--read-uploads-description",
         "y",
@@ -95,7 +95,7 @@ defmodule Mix.Tasks.Pleroma.InstanceTest do
     assert generated_config =~ "http: [ip: {127, 0, 0, 1}, port: 4000]"
 
     assert generated_config =~
-             "filters: [Pleroma.Upload.Filter.Exiftool.StripLocation, Pleroma.Upload.Filter.Exiftool.ReadDescription]"
+             "filters: [Pleroma.Upload.Filter.Exiftool.StripMetadata, Pleroma.Upload.Filter.Exiftool.ReadDescription]"
 
     assert generated_config =~ "base_url: \"https://media.pleroma.social/media\""
     assert File.read!(tmp_path() <> "setup.psql") == generated_setup_psql()